# api/openapi.3.1.projection.yaml
...
openapi: 3.1.0
info:
  title: Bedrock API
  version: 2.0.0
  description: |
    Bedrock 2.0 HTTP API. Envelope `{ code, message, data? }`.
    Pagination data shape: `{ items, total, page, page_size, total_pages }`.
servers:
  - url: /api/v1
    description: Relative to the Bedrock host
security:
  - bearerAuth: []
...
paths:
...
  /build-jobs/{id}/runs:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [BuildJobs]
      summary: Enqueue build run (Wave 3 stub queue; Wave 4 executes)
      description: |
        Requires `cicd.build_jobs:execute` only. Credential `:use` is NOT required at
        trigger time (bound credential snapshot is used at execution).
      operationId: buildJobsEnqueueRun
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                branch: {type: string}
                trigger_type: {type: string}
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/BuildRun"
  /build-runs:
    get:
      tags: [BuildRuns]
      summary: List build runs
      operationId: buildRunsList
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: build_job_id
          in: query
          schema: {type: integer}
        - name: status
          in: query
          schema: {type: string}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/BuildRunPage"
  /build-runs/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [BuildRuns]
      summary: Get build run detail (includes deploy attempts)
      operationId: buildRunsGet
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/BuildRun"
  /build-runs/{id}/cancel:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [BuildRuns]
      summary: Cancel a build run (or cancel in-flight distribution if already success)
      operationId: buildRunsCancel
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/BuildRun"
  /build-runs/{id}/retry:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [BuildRuns]
      summary: Retry by creating a new BuildRun
      operationId: buildRunsRetry
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/BuildRun"
  /build-runs/{id}/redeploy:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [BuildRuns]
      summary: Redeploy artifact on the same BuildRun (appends BuildDeployAttempt batch)
      operationId: buildRunsRedeploy
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                target_ids:
                  type: array
                  items: {type: integer}
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/BuildRun"
  /build-runs/{id}/artifact:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [BuildRuns]
      summary: Download build artifact archive
      operationId: buildRunsArtifact
      responses:
        "200":
          description: Artifact file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /build-runs/{id}/log:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [BuildRuns]
      summary: Fetch build log text (WS preferred for live streaming)
      operationId: buildRunsLog
      responses:
        "200":
          description: Plain text log
          content:
            text/plain:
              schema:
                type: string
...
  /credentials:
    get:
      tags: [Credentials]
      summary: List credentials (metadata only; never plaintext)
      operationId: credentialsList
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: keyword
          in: query
          schema: {type: string}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/CredentialPage"
    post:
      tags: [Credentials]
      summary: Create credential
      operationId: credentialsCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Credential"
  /credentials/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Credentials]
      summary: Get credential metadata
      operationId: credentialsGet
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Credential"
    put:
      tags: [Credentials]
      summary: Update credential (empty secret keeps existing)
      operationId: credentialsUpdate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Credential"
    delete:
      tags: [Credentials]
      summary: Delete credential
      operationId: credentialsDelete
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "409":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    SuccessEnvelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string
        request_id:
          type: string
...
    Page:
      type: object
      required: [items, total, page, page_size, total_pages]
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
          format: int64
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
...
    BuildDeployAttempt:
      type: object
      properties:
        id: {type: integer}
        build_run_id: {type: integer}
        batch_no: {type: integer}
        deploy_target_id: {type: integer, nullable: true}
        target_snapshot_json: {type: string}
        status: {type: string}
        log_path: {type: string}
        error_message: {type: string}
        started_at: {type: string, format: date-time, nullable: true}
        finished_at: {type: string, format: date-time, nullable: true}
        created_at: {type: string, format: date-time}
    BuildRun:
      type: object
      properties:
        id: {type: integer}
        build_job_id: {type: integer}
        build_number: {type: integer}
        status:
          type: string
          enum: [queued, running, success, failed, cancelled, interrupted]
        stage:
          type: string
          enum: [pending, cloning, building, archiving, distributing, idle]
        trigger_type: {type: string}
        triggered_by: {type: integer}
        branch: {type: string}
        commit_hash: {type: string}
        commit_message: {type: string}
        log_path: {type: string}
        artifact_path: {type: string}
        duration_ms: {type: integer, format: int64}
        error_message: {type: string}
        distribution_summary:
          type: string
          enum: [none, running, all_success, partial, all_failed, cancelled]
        snapshot_json: {type: string}
        started_at: {type: string, format: date-time, nullable: true}
        finished_at: {type: string, format: date-time, nullable: true}
        created_at: {type: string, format: date-time}
        deploy_attempts:
          type: array
          items:
            $ref: "#/components/schemas/BuildDeployAttempt"
    BuildRunPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/BuildRun"
...
    Credential:
      type: object
      description: Metadata only; secret never returned
      properties:
        id: {type: integer}
        name: {type: string}
        type: {type: string, enum: [password, token, ssh_key, api_key]}
        username: {type: string}
        description: {type: string}
        has_secret: {type: boolean}
        has_passphrase: {type: boolean}
        created_by: {type: integer}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    CredentialCreateRequest:
      type: object
      required: [name, type, secret]
      properties:
        name: {type: string}
        type: {type: string}
        username: {type: string}
        secret: {type: string}
        passphrase: {type: string}
        description: {type: string}
    CredentialUpdateRequest:
      type: object
      properties:
        name: {type: string}
        type: {type: string}
        username: {type: string}
        secret: {type: string, description: "Empty keeps existing"}
        passphrase: {type: string, description: "Empty keeps existing"}
        description: {type: string}
    CredentialPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Credential"